backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
openai==1.3.5
pandas==2.1.3
openpyxl==3.1.2
pydantic==2.5.2
python-multipart==0.0.6
httpx==0.25.2
thefuzz==0.20.0
python-Levenshtein==0.23.0
pytest==7.4.3
